feature(bento-mega-menu): extracted Shim
