+++ src/preact/types.ts
@@ -16,7 +16,8 @@ export type FC<P = {}> = FunctionalComponent<P>;
 
 export type EventHandler<E extends JSXInternal.TargetedEvent> =
   JSXInternal.EventHandler<E>;
-export type HTMLAttributes = JSXInternal.HTMLAttributes;
+export type HTMLAttributes<RefType extends EventTarget = EventTarget> =
+  JSXInternal.HTMLAttributes<RefType>;
 export type IntrinsicElements = JSXInternal.IntrinsicElements;
 export type PropsWithChildren<TProps = {}> = TProps & {
   children?: ComponentChildren;
